Refactor into applications directory.
